Add paging to audit table entity
