+++ src/ipsis/course-merge/CourseMergeOfferingCollectionEntity.js
@@ -3,9 +3,9 @@
  * See: ISirenCourseMergeSerializer.SerializeCourseOfferingListResult
  */
 import { Actions, Rels } from '../../hypermedia-constants.js';
-import { BaseCollectionEntity } from './BaseCollectionEntity.js';
+import { CourseMergeOfferingBaseCollectionEntity } from './CourseMergeOfferingBaseCollectionEntity.js';
 
-export class CourseMergeOfferingCollectionEntity extends BaseCollectionEntity {
+export class CourseMergeOfferingCollectionEntity extends CourseMergeOfferingBaseCollectionEntity {
 	userOwnedByMultipleSourceSystems() {
 		return this._entity?.properties?.userOwnedByMultipleSourceSystems;
 	}
